feat: export createDefaultOptions [EXT-6172]
